Fix build warnings and improve code style
- Simplify code
- Avoid unnecessary boxing and unboxing
- Use modern language features
- Remove redundant code contructs

+++ src/main/java/com/healthmarketscience/jackcess/impl/expr/DoubleValue.java
@@ -48,7 +48,7 @@ public class DoubleValue extends BaseNumericValue {
 
     @Override
     public boolean getAsBoolean(LocaleContext ctx) {
-        return _val.doubleValue() != 0.0d;
+        return _val != 0.0d;
     }
 
     @Override
